string: implement memmove using memcpy
The current implementation of memmove is a PITA (I mean the bread, of course)
to decode if a fault happens. We have very little control of where exactly in
the code the fault happens, therefore it is difficult to reason about it. This
patch implements memmove in terms of memcpy + memcpybackwards. For those, we
have specific fixups in the possible fault sites, that will allow us to decode
the faults with ease.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>