linux: add priority stubs
Cassandra is crashing for me in some invocations (I really don't understand how
it was running before), because it can't find setpriority and getpriority
symbols.

Since now we have priorities implemented in our scheduler, we can implement the
equivalent functions of linux' getpriority and setpriority. We will ignore all
non-thread entities for now (and possibly forever).

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>