sched: change thread list into a set.
A list can be slow to search for an element if we have many threads. Even under
normal load, the number of threads we span may not be classified as huge, but it
is not tiny either.

Change it to a set so we can implement functions that operate on a given thread
without that much overhead

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>