balloon: do not call the balloon if we have waiters
We have found a balloon related deadlock recently, that goes as follows:

* GC kicks in (effectively blocking all java code to run)
* GC wait()s, because it needs memory, and there is no more memory
* Because there is no more memory, we call the balloon
* But the balloon calls into java/jni, so it blocks, because GC is active

Deadlock.

The only solution to this problem is to not balloon *at all* when there are
waiters. Although we could just restrict that to GC waiters, it is very hard in
practice to pinpoint them, and even "just java waiters" does not seem like a
cost effective measure. Also note, that "call into Java/JNI" is not really a
choice of our design in particular: to balloon we need to allocate an object,
and that's a java call: there is no escape.

As Nadav has once pointed out, balloon is not supposed to be used in
emergencies, and if we have waiters, this is too late. We should have done that
earlier.

In theory, there can be someone doing a very very large allocation that ends up
waiting while GC can safely proceed, and ballooning in that case would help.
However, I guess we will have to just live with that.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>