jvm balloon: does not take the balloon lock during fault
The lock itself is no longer needed - it was in the past before the balloon was
directly put in the vma. But besides that, it is leading to subtle deadlocks:
let's say a new balloon is being inserted into the list, which needs to be done
with lock held. If Garbage Collection kicks in at this very same time and
happens to move a balloon, it will stop in the lock. This means that all java
threads will stop, waiting for the GC to finish. All it takes is a single JNI
call with the lock taken to achieve a deadlock. In particular, deleting a balloon
involves destroying a reference, so that is enough to trigger it.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>