bsd: also use jvm figures in kmem_used calculations
We are defined kmem_used() as total - free. However, as the name implies, this
field is supposed to report on the amount of kernel memory used, not total phys
memory.

Granted, this situation is trickier for us to determine, because we're running
everything together. But in some situations, this can lead to an interesting
performance deterioration: if the JVM touches all the memory for the heap, this
will lead to a very large memory usage, even though the ARC itself is using very
little of it. The ARC will then shrink, and will never have the chance of grow
to the sizes it should. Including the jvm_heap figures will allow the ARC to
have a better idea about what happens.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>