mempool.cc: print the amount of memory available when we oom
We consider ourselves to be "out of memory" when an allocation can not be
satisfied, not when we reach 0 free memory. If we are heavily fragmented, those
two situations can be quite departed from each other. I have never printed the
amount of free memory at oom because abort() could not handle a format. Now it
can.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>