reclaimer: wipe out conditionals
One of the reasons we are being overly aggressive with ZFS reclaim is that once
we reach the reclaimer, unless we take ourselves away from the pressure
situation, we will keep reclaiming until we do. That happens even if we don't
need any extra memory. This is particularly problematic given the way ZFS
reclaimer is designed: if it is called twice in a row, it changes its mode of
operation from conservative (ARC_RECLAIM_CONS) to aggressive
(ARC_RECLAIM_AGGR), and will shrink the cache to the current cache usage.

Since we will wake the reclaimer every time we have reason to believe memory is
under pressure, there is no need for the reclaimer itself to keep looping.

After this patch is applied, the reclaimer is still called sequentially if the
allocation rate is high (because we will wake it many times), but will be left
alone if it isn't.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>