reclaimer: only wake reclaimer when memory is used by OSv itself
The current waking criteria - to wake up the reclaimer when memory is overall
short - won't really work when we have ballooning enabled: in that situation,
the whole of memory will most likely be taken by the heap. This means that we
will constantly be bellow the threshold and will reclaim at every opportunity.

OSv memory usage won't grow, and we won't be placing memory when it is most
needed.  What we really need is to keep waking the reclaimer when memory usage
caused by OSv is high, but for the balloon policy itself we will need our own
calls.

stats::jvm_heap() will return 0 when we have no balloon, therefore that case
will keep seeing no differences.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>