xen blkfront: software barrier support
This code comes largely from BSD's review r252260, with modifications from our infrastructure.
After this patch, filesystem requests that ask for serialization will freeze the request queue
until that command is processed, unless we have Hypervisor support for serialization. That should
be present in any modern Xen hypervisor, but it is not present on EC2.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>