xen: make xenstore functions match the ones for libxenstore
libxenstore functions are named exactly as our internal functions, but the
signatures are different. This means that when trying to run a program that was
written for libxenstore, confusion will ensue.

All our internal users of those functions are already c++-ified, so by just
moving the __END_DECLS, we already get rid of the name conflict, and the names
will resolve to libxenstore. However, libxenstore expects to manage a userpace
xenstore. That means opening sockets, and dealing with multiple things that
would only complicate OSv further.

The best course of action is to provide ourselves the functions with the same
signature libxenstore expects.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>