xen: generate a symbol for event channel notification
We already have a symbol for event channel notification via irq number,
but it would be useful to allow external users access to notifications
via event channels as well. Currently that is done fully inline, which
makes sense in most situations. But let's generate a symbol for that,
so external users can benefit.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>