xenfront: derive unit number from xenstore
We currently bump an integer and have that to be the index number.  However,
the interfaces are already indexed in the xenstore. Following a different
indexing scheme is possible and valid, but has the side effect that referring
to "interface 0" from outside the guest won't necessarily match the same
"interface 0" within it.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>